// frontend/src/api/model/flight.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * DOSPORTAL API
 * API Documentation for DOSPORTAL
 * OpenAPI spec version: 1.0.0
 */
import type { Airport } from './airport';

export interface Flight {
  readonly id: string;
  flight_number: string;
  /** @nullable */
  departure_time?: string | null;
  readonly takeoff: Airport;
  readonly land: Airport;
}